+++ app/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -1893,9 +1893,6 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
    * @return the block period in seconds, or empty if not applicable
    */
   private OptionalInt getBlockPeriodSeconds(final GenesisConfigOptions genesisConfigOptions) {
-    if (genesisConfigOptions.isClique()) {
-      return OptionalInt.of(genesisConfigOptions.getCliqueConfigOptions().getBlockPeriodSeconds());
-    }
     if (genesisConfigOptions.isIbft2()) {
       return OptionalInt.of(genesisConfigOptions.getBftConfigOptions().getBlockPeriodSeconds());
     }
@@ -1916,8 +1913,6 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
       return OptionalLong.of(genesisConfigOptions.getBftConfigOptions().getEpochLength());
     } else if (genesisConfigOptions.isQbft()) {
       return OptionalLong.of(genesisConfigOptions.getQbftConfigOptions().getEpochLength());
-    } else if (genesisConfigOptions.isClique()) {
-      return OptionalLong.of(genesisConfigOptions.getCliqueConfigOptions().getEpochLength());
     }
     return OptionalLong.empty();
   }
@@ -1933,8 +1928,6 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
       return "IBFT2";
     } else if (genesisConfigOptions.isQbft()) {
       return "QBFT";
-    } else if (genesisConfigOptions.isClique()) {
-      return "Clique";
     } else if (genesisConfigOptions.isEthHash()) {
       return "Ethash";
     }